+++ ReleaseNotes/release_notes.md
@@ -15,7 +15,8 @@
   1. further extend `test_trees_params.py` test
   1. add `__bool__` and `__nonzero__` methods for `ProgressBar`  - it allows to make more easy `while`-loops 
   1. change default table layour for `isatty` regime from `SingleTable` to `DoubleTable`
-  
+  1. allow to specify the default table format (`local`,`ascii`,`single`,`double`(default),`porcelain`, `github`,`markdown`
+
 ## Backward incompatible:  
 
   1. rename `BernsteinDualBasis` into `Ostap::Math::BernsteinnDual`